Move `sql_join_builder.py` to `to_sql_plan`.
